Refactors dataset-specific search tests.
